refactor: restructure api controllers
